+++ src/localisation/phrases.ts
@@ -502,6 +502,8 @@ enum Phrase {
   LinuxGsrBufferSecondsDescription,
   LinuxGsrCodecDescription,
   LinuxGsrBitrateKbpsDescription,
+  LinuxGsrAudioOutputDescription,
+  LinuxGsrAudioInputDescription,
   LinuxGsrAudioDescription,
   LinuxGsrReplayStorageDescription,
   LinuxGsrLeadInSecondsDescription,