+++ src/renderer/preload.d.ts
@@ -27,6 +27,11 @@ declare global {
           previewHeight: number;
         }>;
 
+        getLinuxGsrAudioDevices(): Promise<{
+          inputs: Array<{ value: string; label: string }>;
+          outputs: Array<{ value: string; label: string }>;
+        }>;
+
         configurePreview(
           x: number,
           y: number,